Round rather than truncate the reconstructed Monkhorst-Pack grid
cell_find_MP_grid recovers the grid density as 1/min_img, where min_img is the
smallest separation in the time-reversal-completed, folded k-point list. The
k-points come from the .bands file, which CASTEP writes to 8 decimals, so
min_img is a difference of rounded numbers and the reciprocal misses the
integer by a few ulp in either direction. int() truncates, and a miss on the
low side then costs a whole grid point.

1/17 = 0.0588235294... is written as 0.05882353, i.e. rounded up. Every
pairwise difference inherits that, so

    min_img        = 0.05882352999999996
    1.0_dp/min_img = 16.999999830000014

and a 17x17x1 mesh was reported as 16x16x1.

Verified end to end on a Cu(100), one layer, kpoint_mp_grid 17 17 1,
which CASTEP reduces to 45 irreducible k-points. Two binaries differing
only in these two words, same .bands, same .odi -- the .odo files differ
in the "Grid size" line and nothing else.

    -out.cell truth   17  17   1
    int()             16 x 16 x  1
    nint()            17 x 17 x  1

+++ optados/src/cell.f90
@@ -298,14 +298,14 @@ contains
       if (abs(2.0_dp*min_img - min_img2) < min_img_tol) then
         ! If 1stMI==2ndMI then 1/1stMP is the grid density
         if (present(kpoint_offset)) kpoint_offset(idim) = 0.0_dp
-        kpoint_grid_dim(idim) = int(1.0_dp/min_img)
+        kpoint_grid_dim(idim) = nint(1.0_dp/min_img)
         ! WARNING could also have a shifted grid with a perfect shift 3/(4n)
         ! this would be a known bug
       else
         ! If 1stMI.ne.2ndMI then 1/3rdMP is the grid density
         ! and 1stMI/2 is the shift
         if (present(kpoint_offset)) kpoint_offset(idim) = min_img/2.0_dp
-        kpoint_grid_dim(idim) = int(1.0_dp/min_img3)
+        kpoint_grid_dim(idim) = nint(1.0_dp/min_img3)
       end if
 
     end do over_dim
